disable format by default

+++ recipes/drives.rb
@@ -22,11 +22,13 @@ node['rails']['drives'].each do |name, params|
   file_system = params['file_system'] || 'xfs'
   children = params['children'] || false
   mkfs_name = "#{name}#{children ? 1 : ''}"
-  force_format = params['force'] || false
+  force_format = params['force_format'] || false
+  with_format = params['format'] || true
   label = params['label'] || 'loop'
+  mount_point = params['mount_point'] || false
 
   mount_disk = mount(name) do
-    device name
+    device mkfs_name
     options 'defaults,nofail'
     fstype file_system
     action :nothing
@@ -34,14 +36,14 @@ node['rails']['drives'].each do |name, params|
 
   mkfs = execute("mkfs.#{file_system} #{force_format ? '-f ' : ''}#{mkfs_name}") do
     action :nothing
-    notifies :mount, mount_disk, :immediately
-    notifies :enable, mount_disk, :delayed
+    notifies :mount, mount_disk, :immediately if mount_point
+    notifies :enable, mount_disk, :delayed if mount_point
   end
 
   execute "parted #{name} --script -- mklabel #{label} mkpart #{part_type} #{file_system} 1 -1s" do
     # Number  Start   End    Size   File system  Name  Flags
     #  1      17.4kB  537GB  537GB               xfs
     not_if "parted #{name} --script -- print |sed '1,/^Number/d' |grep #{part_type}"
-    notifies :run, mkfs, :immediately
+    notifies :run, mkfs, :immediately if with_format
   end
 end
